+++ services/Settings/Definition/unit.service.js
@@ -1,6 +1,7 @@
 const unitModel = require("../../../models/Settings/Definition/unit.model");
 const ApiError = require("../../../utils/apiError");
 // const productModel = require("../../../models/Stocks/Products/product.model");
+const productModel = require("../../../models/productModel");
 
 exports.getUnits = async ({ companyId }) => {
   const query = { companyId };
@@ -42,19 +43,19 @@ exports.deleteUnit = async ({ companyId, id, session }) => {
     throw new ApiError(`No Unit for this id ${id}`, 404);
   }
 
-  // const unitUsed = await productModel
-  //   .exists({
-  //     companyId,
-  //     $or: [{ unit: id }, { "unitsPrices.unitId": id }],
-  //   })
-  //   .session(session);
+  const unitUsed = await productModel
+    .exists({
+      companyId,
+      $or: [{ unit: id }, { "unitsPrices.unitId": id }],
+    })
+    .session(session);
 
-  // if (unitUsed) {
-  //   throw new ApiError(
-  //     `Cannot delete unit because it is linked to product`,
-  //     404,
-  //   );
-  // }
+  if (unitUsed) {
+    throw new ApiError(
+      `Cannot delete unit because it is linked to product`,
+      404,
+    );
+  }
 
   await unit.deleteOne({ session });
